+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.6] Unreleased
+
+### Added
+- methods and notification related to holdinvoices
+- new `state` field in transaction responses
+- include pending and failed payments in ``list_transactions``
+- include expired invoices in ``list_transactions`` unless `unpaid` is set
+- missing fields for ``make_invoice`` response
+- new `fees_paid` field in ``pay_keysend``/``multi_pay_keysend``
+
+### Fixed
+- Don't ignore `offset` parameter in ``list_transactions``
+
 ## [0.1.5] 2025-07-24
 
 ### Changed